requirements.txt
PyGithub
openai
numpy
plantuml
requests